requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-classic
langchain-groq
langgraph
chromadb
pypdf
python-dotenv
sentence-transformers
tavily-python